Update README.md
Random Number Guessing Game

Write a program that generates a random number and asks the user to guess what the number is. If the user’s guess is higher than the random number, the program should display “Too high, try again.” If the user’s guess is lower than the random number, the program should display “Too low, try again.” The program should use a loop that repeats until the user correctly guesses the random number.

The program should also keep a count of the number of guesses that the user makes. When the user correctly guesses the random number, the program should display the number of guesses.

You can assume a reasonable range for the number (like 1-10).